devtools/client/memory/components/test/mochitest/test_tree_02.html

<!DOCTYPE HTML>
<html>
<!--
Test that collapsed subtrees aren't rendered.
-->
<head>
  <meta charset="utf-8">
  <title>Tree component test</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript;version=1.8"></script>
<script type="application/javascript;version=1.8">
window.onload = Task.async(function* () {
  try {
    let ReactDOM = bRequire("devtools/client/shared/vendor/react-dom");
    let React = bRequire("devtools/client/shared/vendor/react");
    let Tree = React.createFactory(bRequire("devtools/client/memory/components/tree"));

    const tree = ReactDOM.render(Tree(TEST_TREE_INTERFACE), window.document.body);

    yield setState(tree, {
      expanded: new Set("MNO".split(""))
    });

    isRenderedTree(document.body.textContent, [
      "A:false",
      "M:false",
      "-N:false",
      "--O:false",
    ], "Collapsed subtrees shouldn't be rendered");
  } catch(e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }
});
</script>
</pre>
</body>
</html>
